# backoffice/docs/backoffice.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Banking Backoffice",
    "version": "3.0.0"
  },
  "tags": [
    {
      "name": "UserService"
    }
  ],
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/account/create": {
      "post": {
        "summary": "Create New Account",
        "operationId": "UserService_CreateAccount",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CreateAccountRes"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/v1ErrorMessage"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CreateAccountReq"
            }
          },
          {
            "name": "X-API-Key",
            "description": "API Key For Using Backoffice",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "Backoffice Account Management"
        ]
      }
    },
    "/user/create": {
      "post": {
        "summary": "Create New User",
        "operationId": "UserService_CreateUser",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CreateUserRes"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/v1ErrorMessage"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CreateUserReq"
            }
          },
          {
            "name": "X-API-Key",
            "description": "API Key For Using Backoffice",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "Backoffice User Management"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1CreateAccountReq": {
      "type": "object",
      "properties": {
        "userID": {
          "type": "string"
        }
      }
    },
    "v1CreateAccountRes": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "accountID": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "v1CreateUserReq": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        }
      }
    },
    "v1CreateUserRes": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "userID": {
          "type": "string"
        }
      }
    },
    "v1ErrorMessage": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int64"
        },
        "message": {
          "type": "string"
        }
      },
      "title": "common models"
    }
  }
}